+++ packages/gatsby-plugin-theme-ui/test/provider.js
@@ -23,6 +23,7 @@ test('renders with theme context', () => {
   const root = render(wrapRootElement({ element: <Consumer /> }, {}))
   expect(context.theme).toEqual({
     colors: {},
+    rawColors: {},
   })
 })